Updated Deprecated getLatestResult for photonvision
GetLatestResult now returns an optional type, which could contain the value or null if their are no new values, to prevent outOfBoundsExceptions. if there is no new results, we do not estimate the pose, which has a photonvision update method which would catch repeated entries and stop the process after the pose is estimated, returning an empty pose estimate. TLDR: Checks and stops one step earlier if there are no new results.